move elastic search controller
