+++ app/src/main/java/com/example/loggerdoc/elasticclient/ElasticCacheClient.java
@@ -1,30 +1,15 @@
 package com.example.loggerdoc.elasticclient;
 
 import android.content.Context;
-import android.os.AsyncTask;
-
-import com.example.loggerdoc.CareGiver;
-import com.example.loggerdoc.ElasticSearchController;
-import com.example.loggerdoc.Patient;
-import com.example.loggerdoc.User;
-import com.example.loggerdoc.UserList;
-import com.google.gson.Gson;
-import com.google.gson.JsonObject;
-
-import org.json.JSONArray;
-import org.json.JSONException;
-import org.json.JSONObject;
 
 import java.io.BufferedReader;
 import java.io.BufferedWriter;
 import java.io.File;
-import java.io.FileNotFoundException;
 import java.io.FileOutputStream;
 import java.io.FileReader;
 import java.io.IOException;
 import java.io.OutputStream;
 import java.io.OutputStreamWriter;
-import java.util.ArrayList;
 import java.util.concurrent.locks.ReentrantLock;
 
 import okhttp3.OkHttpClient;
