+++ app/src/main/java/com/example/loggerdoc/elasticclient/getProblemsTask.java
@@ -2,9 +2,7 @@ package com.example.loggerdoc.elasticclient;
 
 import android.content.Context;
 import android.os.AsyncTask;
-import android.util.Log;
 
-import com.example.loggerdoc.ElasticSearchController;
 import com.example.loggerdoc.Problem;
 import com.google.gson.Gson;
 import org.json.JSONArray;
