+++ app/src/main/java/com/example/loggerdoc/elasticclient/modifyUserTask.java
@@ -3,7 +3,6 @@ package com.example.loggerdoc.elasticclient;
 import android.content.Context;
 import android.os.AsyncTask;
 
-import com.example.loggerdoc.ElasticSearchController;
 import com.example.loggerdoc.User;
 import com.google.gson.Gson;
