+++ app/src/main/java/com/example/loggerdoc/elasticclient/removeProblemTask.java
@@ -3,11 +3,9 @@ package com.example.loggerdoc.elasticclient;
 import android.content.Context;
 import android.os.AsyncTask;
 
-import com.example.loggerdoc.ElasticSearchController;
 import com.example.loggerdoc.Problem;
 import com.example.loggerdoc.ProblemRecordListController;
 import com.google.gson.Gson;
-import com.google.gson.GsonBuilder;
 
 import java.io.BufferedWriter;
 import java.io.File;
